Revamp MainSidebarMenuItem
The component now takes an href and onClick handler and directly
renders the <a> required by the template.

The text of the menu item has
been moved into a title property so that it can be wrapped in a span
which is required.

An iconName property is used to display a fontawesome icon if specified.

AddonList and Addon components are used to add labels to the right side
of the menu item and are passed in as children.